requirements.txt

duden
pronouncing
discord.py
wikipedia
pytz
